requirements.txt
groq==0.5.0
python-dotenv==1.0.1
Flask==2.3.2
Flask-CORS==4.0.0